// src/enums/fashion-insights-enum.ts
export enum FashionInsightDataType {
  CELEBRITY_STYLE_DATA = "celebrityStyleData",
  EXISTING_WARDROBE = "existingWardrobe",
  WAYS_TO_DRESS_LIKE = "waysToDressLike",
  CAPSULE_PROFESSION = "capsuleProfession",
  COMPARISON_REPORT = "comparisonReport",
  ACCESSORY_RECOMMENDATION = "accessoryRecommendation",
  ITEM_RECOMMENDATION = "itemRecommendation",
  OUTFIT_RECOMMENDATION_VACATION = "outfitRecommendationVacation",
  OUTFIT_RECOMMENDATION_OCCASION = "outfitRecommendationOccasion",
}